A terrible, horrible, working proof-of-concept.
A terrible, horrible, working proof-of-concept.  This needs to be
cleaned up and abstracted out so that it can easily be reused and
extended, especially in the extremely likely event that the template
won't remain as it currently is forever.

Some future items:
* Move away from winforms.  This doesn't need to be a GUI application at
all, it should run as a scheduled job somewhere.
* Set the default url to the most recent mailchimp newsletter in the
mailchimp RSS feed for Mercy Hill.
* Email the resulting snapshots to a mailing list.